hidden, dedup search results
A hidden file or directory doesn’t show in a directory index or in search results. Currently there is no way of syncing this data from Google, but it can be accomplished by manually editing `var/index.json` to add a `”hidden”:true` property to the `index` property of the file in question.

Next step is probably just to have this be part of the saved config.